Move selenium test to `TestWebBrowser` class
